SalaryLens-Predictor-de-Retorno-Laboral-con-Machine-Learning/salarlens_predictor/config.py
from pathlib import Path

from dotenv import load_dotenv
from loguru import logger

# Load environment variables from .env file if it exists
load_dotenv()

# Paths
PROJ_ROOT = Path(__file__).resolve().parents[1]
logger.info(f"PROJ_ROOT path is: {PROJ_ROOT}")

DATA_DIR = PROJ_ROOT / "data"
RAW_DATA_DIR = DATA_DIR / "raw"
INTERIM_DATA_DIR = DATA_DIR / "interim"
PROCESSED_DATA_DIR = DATA_DIR / "processed"
EXTERNAL_DATA_DIR = DATA_DIR / "external"

MODELS_DIR = PROJ_ROOT / "models"

REPORTS_DIR = PROJ_ROOT / "reports"
FIGURES_DIR = REPORTS_DIR / "figures"

# If tqdm is installed, configure loguru with tqdm.write
# https://github.com/Delgan/loguru/issues/135
try:
    from tqdm import tqdm

    logger.remove(0)
    logger.add(lambda msg: tqdm.write(msg, end=""), colorize=True)
except ModuleNotFoundError:
    pass


# src/config.py

# Columnas numericas (continuas o discretas)
NUMERIC_COLS = [
    'experience_years',
    'skills_count',
    'certifications',
    'salary'
]

# Columnas categoricas nominales (sin orden intrinseco)
CATEGORICAL_COLS = [
    'job_title',
    'industry',
    'location',
]

# Columnas categoricas ordinales (con un orden logico)
ORDINAL_COLS = [
    'education_level',
    'company_size',
    'remote_work',
]
