SalaryLens-Predictor-de-Retorno-Laboral-con-Machine-Learning/requirements.txt
ipython
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pandas
pip
pytest
python-dotenv
ruff
scikit-learn
tqdm
typer
seaborn
ipykernel
ydata-profiling
setuptools
ipyidgets
-e .
